feat(#S11P31D101-416): 수업/숙제/시험 관련 레포지토리 구현

+++ backend/lecture-service/src/main/java/com/eum/lecture_service/command/repository/exam/ExamRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.eum.lecture_service.command.entity.exam.Exam;
 
-public interface ExamRepository extends JpaRepository<Exam, Integer> {
+public interface ExamRepository extends JpaRepository<Exam, Long> {
 }
